Started refactoring the auth component.
